requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
dj-rest-auth==7.0.1
Django==5.2.9
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
mysqlclient==2.2.7
pillow==12.1.0
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.5
stripe==14.2.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
validate-email-address==1
